+++ ft_print_u.c
@@ -21,12 +21,12 @@ int	ft_print_u(unsigned int n, int base)
 	count = 0;
 	if (n == 0)
 		return (count = write (1, "0", 1));
-	while (n >= base)
+	while (n >= (unsigned int)base)
 	{
-		count += ft_print_u(n / base, base);
+		count += ft_print_u(n / (unsigned int)base, base);
 		n = n % base;
 	}
-	if (n < base)
+	if (n < (unsigned int)base)
 	{
 		write(1, &basesymb[n], 1);
 		count++;
